% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdna_plan_printable.R
\name{cdna_plan_printable}
\alias{cdna_plan_printable}
\title{cdna_plan_printable}
\description{
Simple script, takes the output from rna2cdna and then makes it an easy to print form for taking ito lab
}
