% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_options}
\alias{color_options}
\title{The different colors used for the COK palette}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{hue_1}{The first color that should be used}
  \item{hue_2}{The second color that should be used}
  \item{hue_3}{The third color that should be used}
  \item{hue_4}{The fourth color that should be used}
  \item{hue_5}{The fifth color that should be used}
  \item{hue_6}{The sixth color that should be used}
}}
\usage{
color_options
}
\description{
A dataset containing the color codes
}
\keyword{datasets}
