% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_categories.R
\name{column_categories}
\alias{column_categories}
\title{Create categories for the column content.}
\usage{
column_categories(
  df,
  col_in,
  col_out,
  quantiles = c(0.25, 0.75),
  labels = c("1", "2", "3"),
  NA_as_zero = TRUE,
  favour_lowest_score = FALSE
)
}
\arguments{
\item{df}{input data}

\item{col_in}{input column, typically "score" as final score}

\item{col_out}{output column, typically "cat_XXX" for the categories}

\item{quantiles}{limits in % of the data to break the col_in into
categories, default c(0.25, .5, .75)}

\item{labels}{string vector for the values between the quantiles,
default c("1", "2", "3")}

\item{NA_as_zero}{bool, default TRUE, to convert the NA to "0"}

\item{favour_lowest_score}{bool, default FALSE, if TRUE the intervals are closed on the left. This is important when the quantiles are very bad, and the cut function does not work well}
}
\description{
The quantiles define the boundaries to make groups
}
\details{
The value "0" for the score 0 is taken automatically

Using default quantiles: c(0.25, .75) and labels c("1", "2", "3") will
make a category = "1" for all values that are between 0 and first quartile,
category = "2" for all values between 0.25 and 0.75 and 3 for values > 0.75

NA_as_zero = TRUE by default, so all zeros and NA will be in category "0"
}
\examples{
\dontrun{
tibble(medic_id = c(0, 0, 1, 2, 3), score_pub = c(0, NA, 0.1, 0.5, 1)) \%>\%
column_categories(score_pub, cat_publi, quantiles = c(0.25, .75),
labels = c("1", "2", "3"), NA_as_zero = T) \%>\%
column_categories(score_pub, cat_publi_2, quantiles = c(0.25, .75),
labels = c("1", "2", "3"), NA_as_zero = F)
}
}
