% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect DB}
\usage{
connect_db(config_path, credential_name, set_utf8 = TRUE)
}
\arguments{
\item{config_path}{config file path, e.g. "~/dbconfig.yml" ; file.edit("~/dbconfig.yml")}

\item{credential_name}{credential name, e.g. "aws_devel"}

\item{set_utf8}{logical, whether to set utf8}
}
\value{
a connection handle
}
\description{
A simple function that return a connection handle, using a config file yaml
}
