% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddct.R
\name{ddct}
\alias{ddct}
\title{ddct}
\arguments{
\item{ct_data_table}{Pre-prcoessed data here. Needs following minimum columns: "group" (100% necessary), "ct_gene" (can be named anything), and "ct_18S" (can specify name)}

\item{refgroup}{specfiy which group within the treatment_group column is your reference group (default = "control")}

\item{targetgene}{Specify EXACTLY name of target gene in the ct_data_table}

\item{refgene}{Specify EXACTLY name of refence gene in the ct_data_table (default "ct-18s")}

\item{refgene.cutoff}{CT cutoff for referene gene (default = 15)}
}
\description{
Taq it easy! Here's a script that takes your CT values for your 18S and gene of interest
and does all those sweet ddCT and fold change calcs for you
}
