% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddct-sybr.R
\name{ddct_sybr}
\alias{ddct_sybr}
\title{ddct_sybr}
\arguments{
\item{input_cts}{This is the}

\item{ref_group}{This is the EXACT name of the treatment_group that is being used as a control group for calcluating ddct (i.e. change from this group)}

\item{cutoff_18S}{This is the cutoff over which 18S values will be discarded (default = 15)}
}
\description{
Workflow for calcuating differential gene expression from quantstudio output using Sybr green reagents.

Input must be in this format for this script to work easily

Column 1: sample_id
Column2: treatment_group
Column3 - x: Cts for various genes of interest
Last Column: CT for 18S (ideally called "ct_18s")
}
