% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddct-taq.R
\name{ddct_taq}
\alias{ddct_taq}
\title{ddct_taq}
\arguments{
\item{quantstudio_output}{This is the 'results" page of the quantstudio output (currently requires to be exactly just at the results - to fix up and make it so just improt whole sheet methinks)}

\item{plate_map}{#this is a long list of which sample is in which well A1, A2, A3 etc (use output form melt384 for easiness perhaps)}

\item{refgene}{Specify EXACTLY what you're refence gene is called in the Quantstdio import (default VIC-18S)}
}
\description{
Taq it easy! Here's a script that takes the Quantstudio RT-PCR output and your PCR template and then will go
and make it all good man and do all that sweet ddCT calcs for you (TODO update)
}
