% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_periods.R
\name{expand_periods}
\alias{expand_periods}
\title{Expand periods}
\usage{
expand_periods(
  df,
  id_col = "id",
  start_col = "start",
  end_col = "end",
  out_col = "day",
  range_of = "day"
)
}
\arguments{
\item{df}{data frame}

\item{id_col}{column name of the id}

\item{start_col}{column name of the start date (use as_datetime)}

\item{end_col}{column name of the end date}

\item{out_col}{column name of the output}

\item{range_of}{range to use to create the sequence. "day" by default}
}
\value{
data frame
}
\description{
From a data frame with start and end date, this function create
a new data frame with a range of days between start and end.
}
\details{
Steps:
  extract each start and end of periods
  create a range of days (see range_of)
  add the ID, rename output column, and join with the original data
}
