% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_meeting_ICS.R
\name{export_meeting_ICS}
\alias{export_meeting_ICS}
\title{Create a minimal iCalendar VEVENT
Using the lubridate::with_tz() function convert in the expected timezone
using R local, see Sys.timezone()}
\usage{
export_meeting_ICS(start, end, meeting_title, outfile = NULL, domain = NULL)
}
\arguments{
\item{start, end}{start and end times of the event, can be a list to create
a serie. This will be converted to GMT time zone (expected by iCalendar)}

\item{meeting_title}{a meeting title of the event.}

\item{outfile}{output file name. If NULL, the iCalendar is returned as a string}

\item{domain}{a domain name. Concatenated to the UUID if given}
}
\value{
text file with the iCalendar format
}
\description{
TODO: add LOCATION, DESCRIPTION, ATTENDEE, ORGANIZER, etc.
}
\examples{
dates <- c("2024-09-14", "2024-09-21", "2024-09-28")
starts <- paste0(dates, paste0(" 10:00:00 ", Sys.timezone()))
ends <- paste0(dates, paste0(" 11:30:00 ", Sys.timezone()))
export_meeting_ICS(starts, ends, "Shopping")

}
