% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sliced_excel.R
\name{export_sliced_excel}
\alias{export_sliced_excel}
\title{Split a dataframe into parts of size X}
\usage{
export_sliced_excel(.df, path_prefix = "./file", slice_size = 10000)
}
\arguments{
\item{.df}{dataframe to export}

\item{path_prefix}{path to the output file default: ./file suffix _partX.xlsx}

\item{slice_size}{slices size for the files}
}
\description{
Split a dataframe into parts of size X
}
