% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json.R
\name{extract_json}
\alias{extract_json}
\title{Extract json from column}
\usage{
extract_json(.x, .id, .col, in_parallel = F, workers = 12, relationship = NULL)
}
\arguments{
\item{.x}{dataframe tile tibble()}

\item{.id}{column id to use to extract data properly}

\item{.col}{column containing the json}

\item{in_parallel}{if TRUE, use furrr::future_map2_dfr instead of purrr::map2_dfr}

\item{workers}{number of workers to use, default 12}

\item{relationship}{if not NULL, you probalby need "many-to-many"}
}
\description{
Extract json from column
}
