% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json_v2.R
\name{extract_json_v2}
\alias{extract_json_v2}
\title{Extract json from column}
\usage{
extract_json_v2(df, col, json_type = "array", names_repair = "check_unique")
}
\arguments{
\item{df}{dataframe}

\item{col}{column name to extract. Should not be a string, but a dplyr-column name}

\item{json_type}{either "array" or "object". In case of "name_repair" error,
first check that you do not have already a column with the same name as the
one you will create (ex: "name already exists" for any entity name)
Duplicated names will be repaired when the optin name_repair is used}

\item{names_repair}{option passed to unnest, check ?tidyr::unnest.
Advice: use "unique" or "universal" to repair, default to "check_unique"}
}
\description{
Extract json from column
}
