% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json_v3.R
\name{extract_json_v3}
\alias{extract_json_v3}
\title{Extract json from column simple version}
\usage{
extract_json_v3(
  df,
  .col = "json_data",
  .id = "id",
  if_error = tibble(),
  filter_json = ""
)
}
\arguments{
\item{df}{dataframe with a column json_data and a column id}

\item{.col}{column containing the json, default "json_data"}

\item{.id}{column containing the id, default "id"}

\item{if_error}{dataframe to return in case of error, default tibble()}

\item{filter_json}{filter to apply on the json after extraction, for
example when there is a list and we want only the "primary == TRUE"}
}
\value{
dataframe with json extracted
}
\description{
Extract json from column simple version
}
