% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json_v5.R
\name{extract_json_v5}
\alias{extract_json_v5}
\title{Extract json from column}
\usage{
extract_json_v5(.df, .col)
}
\arguments{
\item{.df}{dataframe}

\item{.col}{column containing the json, default "properties"}
}
\description{
Extract json from column
}
