% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt384.R
\name{melt384}
\alias{melt384}
\title{Take a 384 well plate output and melt to tidy data}
\usage{
melt384(x, y = 3)
}
\arguments{
\item{x}{this is the first number to input}

\item{y}{this is the second number to input}
}
\description{
Testing first as a simple function (takes 2 inputs)
}
