% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt96.R
\name{melt96}
\alias{melt96}
\title{melt96}
\usage{
melt96(samples)
}
\arguments{
\item{samples}{this is your 96 well plate data (8 x 12 format)}
}
\description{
This script takes an input of a 384 well template and melts it to give the value with well position
}
\examples{
samples <- read.csv(my_96_well_plate.csv)
samples_long_list <- handytools::melt96(samples)

# Example output:
#  well_position  sample_id
#      A1         220
#      A2         221
#      A3         222 , etc....

}
