% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_0_to_1.R
\name{normalize_0_to_1}
\alias{normalize_0_to_1}
\title{Normalize values from any range to a range between 0 and 1}
\usage{
normalize_0_to_1(x, na.rm = TRUE, all_same_return = 0.5)
}
\arguments{
\item{x}{vector of numerical values, or column name from a piped data}

\item{na.rm}{option to consider the NA or not, default TRUE}

\item{all_same_return}{control over value to return when all same
BBmisc::normalize return 0.5, but in some cases we want to return 1}
}
\value{
vector
}
\description{
Normalize values from any range to a range between 0 and 1
}
\examples{
normalize_0_to_1(c(0, 1.87, 98, 1, 5, -50, -25))
\dontrun{
data.frame(x = c(0, 1.87, 98, 1, 5, -50, -25)) \%>\% 
mutate(y = normalize_0_to_1(x))
}
}
