% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_na.R
\name{paste_na}
\alias{paste_na}
\title{paste_na}
\usage{
paste_na(x, sep = ",")
}
\arguments{
\item{x}{a vector}

\item{sep}{default ','}
}
\value{
string
}
\description{
A function that paste data, similar to paste or paste0,
but avoid to have NA as string when the data is actually missing
}
