% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Produces a plot of the color palette}
\usage{
plot_colors(df)
}
\arguments{
\item{df}{Dataframe of dimension 10 x 6}
}
\value{
Plot
}
\description{
Produces a plot of the color palette
}
