% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismify.R
\name{prismify}
\alias{prismify}
\title{prismify}
\description{
Takes a tidy dataset with two columns (column 1: group, column 2: value) and then arranges so each group is it's own column, to make it easy to put into prism.
}
