% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_scrape.R
\name{safe_scrape}
\alias{safe_scrape}
\title{safe-scrape}
\usage{
safe_scrape(cmd, if_error = NA_character_)
}
\arguments{
\item{cmd}{an expression to be evaluated. Use curvy brackets}

\item{if_error}{value to return if error occurs}
}
\value{
depends on the cmd, generally a string
}
\description{
A safe wrapper for scraping functions
}
