% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env_variable.R
\name{set_env_variable}
\alias{set_env_variable}
\title{Set an environment variable
Set an environment variable. If the variable already exists,
it will not be overwritten unless the overwrite argument is set to TRUE.}
\usage{
set_env_variable(key, value, overwrite = FALSE)
}
\arguments{
\item{key}{The name of the environment variable to set}

\item{value}{The value of the environment variable to set}

\item{overwrite}{If TRUE, environment variable will be overwritten}
}
\value{
boolean
}
\description{
Set an environment variable
Set an environment variable. If the variable already exists,
it will not be overwritten unless the overwrite argument is set to TRUE.
}
\examples{
set_env_variable("API_KEY", "your_api_key_here")
Sys.getenv("API_KEY")
Sys.unsetenv("API_KEY")
}
