% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_convert.R
\name{uuid_convert}
\alias{uuid_convert}
\title{convertion tool for the uuid}
\usage{
uuid_convert(id, hyphens = TRUE)
}
\arguments{
\item{id}{UUID (strings) to convert}

\item{hyphens}{logical, whether to include hyphens in the output (default: TRUE)}
}
\value{
A string representation of the UUID, with or without hyphens
}
\description{
convertion tool for the uuid
}
