% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect DB}
\usage{
connect_db(
  credential_name = NULL,
  config_path = "~/dbconfig.yml",
  command = "set names utf8"
)
}
\arguments{
\item{credential_name}{credential name, e.g. "aws_devel"}

\item{config_path}{config file path, e.g. "~/dbconfig.yml" ; file.edit("~/dbconfig.yml")}

\item{command}{command to run after connect, typical: set utf8}
}
\value{
a connection handle
}
\description{
A simple function that return a connection handle, using a config file yaml
}
