% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json_v6.R
\name{extract_json_v6}
\alias{extract_json_v6}
\title{Extract json from column}
\usage{
extract_json_v6(.x, .id, .col, relationship = NULL, rm_col = TRUE)
}
\arguments{
\item{.x}{dataframe}

\item{.id}{the ID column to use in case of line duplication}

\item{.col}{column with json to extract}

\item{relationship}{if some many-to-many are expected, use this option}

\item{rm_col}{remove the column from the df returned since data is extracted}
}
\value{
dataframe with extracted json columns#'
}
\description{
Extract json from column
}
