% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt384.R
\name{melt384}
\alias{melt384}
\title{melt384}
\usage{
melt384(samples)
}
\arguments{
\item{x}{this is the first number to input}

\item{y}{this is the second number to input}
}
\description{
melt384
}
\examples{
samples <- any_384_well_dataframe # needs to be 24 x 16
melt384(samples)
output:
     A1  220

}
