% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt384.R
\name{melt384}
\alias{melt384}
\title{melt384}
\usage{
melt384(samples)
}
\arguments{
\item{samples}{this is your 384 well plate data (16 x 24 format)}
}
\description{
melt384
}
\examples{
samples <- read.csv(my__384_well_plate)   #needs to be
samples_tidy <- handytools::melt384(samples)

# Example output:
#   position   value
#      A1      220
#      A2      221
#      A3      222 , etc....

}
